# Please enter the commit message for your changes. Lines starting # with '#' will be ignored, and an empty message aborts the commit. # On branch master # Changes to be committed: #   (use "git reset HEAD <file>..." to unstage) # #	modified:   include/simplefluid.h #	modified:   include/sparsematrix.h #	modified:   src/simplefluid.cpp # # Changes not staged for commit: #   (use "git add <file>..." to update what will be committed) #   (use "git checkout -- <file>..." to discard changes in working directory) # #	modified:   src/main.cpp # # Untracked files: #   (use "git add <file>..." to include in what will be committed) # #	include/cgsolver.h #	res/ #	src/cgsolver.cpp
Fixing some errors in simple fluid and sparse matrix

Forgot to add some of the template parameters in the matrix market file loader for SparseMatrix, this has been added. Also due to some issues with cg solving an int matrix I've switched the SimpleFluid's interaction matrix to just be a float matrix instead

+++ include/simplefluid.h
@@ -32,7 +32,7 @@ private:
 	* Generate the cell-cell interaction matrix for this simulation
 	* where diagonal entries are 4 and neighbor cells are -1
 	*/
-	SparseMatrix<int> createInteractionMatrix();
+	SparseMatrix<float> createInteractionMatrix();
 	/*
 	* Compute the cell number of a cell at the x,y coordinates
 	*/
@@ -47,7 +47,7 @@ private:
 	tcl::Context context;
 	int dim;
 	Window &window;
-	SparseMatrix<int> interactionMat;
+	SparseMatrix<float> interactionMat;
 	//The quad to be used for drawing the fluid texture
 	const static std::array<glm::vec3, 8> quad;
 	//The quad's element buffer
